requirements.txt
psycopg2-binary==2.9.6
Django==4.1.5
whitenoise==6.5.0
python-dotenv==1.0.0
djangorestframework
debugpy
parameterized
